# dqMetadataSpecificationApi/target/classes/spec/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Credit Statement Services US",
        "description": "Services to retrieve credit account statements",
        "contact": {
            "email": "DL-PayPalCredit-Altus-Pilot@paypal.com"
        }
    },
    "schemes": [
        "http"
    ],
    "basePath": "/v1/ppcstatementserv",
    "tags": [
        {
            "name": "CreditStatements",
            "description": "Credit Statement Services",
            "externalDocs": {
                "description": "Find out more",
                "url": ""
            }
        }
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/accounts/{account_id}/statements": {
            "get": {
                "summary": "Returns a list of statement headers for an account.",
                "tags": [
                    "CreditStatements"
                ],
                "description": "List of statement headers for the past 24 months for an account. Statement header includes info such as statement date, balance amount and available media types of the statement",
                "x-api-name": "StatementsResource",
                "operationId": "getStatements",
                "parameters": [
                    {
                        "name": "account_id",
                        "in": "path",
                        "description": "The customer account ID.",
                        "type": "string",
                        "format": "credit_account_id_v1",
                        "minLength": 16,
                        "maxLength": 16,
                        "pattern": "^[0-9]{16}$",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "./statement_header.json"
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "./openapi_common_components_v3/error.json"
                        }
                    },
                    "400": {
                        "description": "Invalid Account id supplied",
                        "schema": {
                            "$ref": "./openapi_common_components_v3/error.json"
                        }
                    },
                    "404": {
                        "description": "Account or Statement with given id not found"
                    }
                }
            }
        },
        "/accounts/{account_id}/statements_with_request_context": {
            "get": {
                "summary": "Testing request context vendor extension",
                "tags": [
                    "CreditStatements"
                ],
                "description": "List of statement headers for the past 24 months for an account. Statement header includes info such as statement date, balance amount and available media types of the statement",
                "x-api-name": "StatementsResource",
                "x-codegen-server-request-context": true,
                "operationId": "getStatementsWithRequestContext",
                "parameters": [
                    {
                        "name": "account_id",
                        "in": "path",
                        "description": "The customer account ID.",
                        "type": "string",
                        "format": "credit_account_id_v1",
                        "minLength": 16,
                        "maxLength": 16,
                        "pattern": "^[0-9]{16}$",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "./statement_header.json"
                            }
                        }
                    },
                    "404": {
                        "description": "Account or Statement with given id not found"
                    }
                }
            }
        },
        "/accounts/{account_id}/statements_with_async_response": {
            "get": {
                "summary": "Testing request context vendor extension",
                "tags": [
                    "CreditStatements"
                ],
                "description": "List of statement headers for the past 24 months for an account. Statement header includes info such as statement date, balance amount and available media types of the statement",
                "x-api-name": "StatementsResource",
                "x-codegen-server-async": true,
                "operationId": "getStatementsWithAsyncResponse",
                "parameters": [
                    {
                        "name": "account_id",
                        "in": "path",
                        "description": "The customer account ID.",
                        "type": "string",
                        "format": "credit_account_id_v1",
                        "minLength": 16,
                        "maxLength": 16,
                        "pattern": "^[0-9]{16}$",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "./statement_header.json"
                            }
                        }
                    },
                    "404": {
                        "description": "Account or Statement with given id not found"
                    }
                }
            }
        }
    }
}